Move current codegen to `codegen/` subdir
