+++ src/app/clusters/power-source-server/app_config_dependent_sources.cmake
@@ -16,5 +16,5 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/power-source-server.cpp"
+    "${CLUSTER_DIR}/codegen/power-source-server.cpp"
 )
