+++ src/app/clusters/power-source-server/app_config_dependent_sources.gni
@@ -11,4 +11,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-app_config_dependent_sources = [ "power-source-server.cpp" ]
+app_config_dependent_sources = [ "codegen/power-source-server.cpp" ]
